webapp/requirements.txt
starlette
databases
sqlalchemy
aiofiles
jinja2
bootstrap4
python-multipart
azure-storage-blob
python-dotenv
asgi-auth-github
# Required for blob storage
aiohttp
# Required for video indexer
aiohttp[speedups]
# Required for deploying to Azure
uvicorn
uvloop
httptools